fix: add a flag for recursive checksum generation and disable by default
When we run bundle install we generate a Gemfile.lock. When we run bundle install --deployment we install all gems into the vendor/bundle folder (and these gems can include Gemfile.lock files inside them). With recursive search for Gemfile.lock we add these new files to the checksum when we save the cache resulting in misses next restore.

https://linear.app/buildkite/issue/MDC-550