fix: allow absolute paths which are under $HOME and increase testing
Currently zstash relies on paths either being under $HOME or within the working directory of a project. This is done to make things easier when moving between machines with different $HOME paths.

That said this behavour was only working if the cache path provided was using ~/ prefix, and not working for absolute paths which were in fact under the $HOME directory.

I have resolved this and a few other small pain points based on feedback.