feat: adds clean before restore to avoid merging data
This change provides cleaning of paths before restore which is critical in ensuring that the resulting path is exactly what is in the cache archive.

This routine has checks to avoid remove of `/`, `~/` or `C:\` paths.

It also caters for Go Modules caches which are set to readonly by the Go command, these are first made writable before deletion.